chore(opflags): change leech deadline to 10 minutes

+++ web/router.go
@@ -28,7 +28,7 @@ const (
 	InfohashKey     InfohashKeyType = "infohash"
 )
 
-func StartWeb(address string, timeout int64, cred map[string][]byte, db persistence.Database) {
+func StartWeb(address string, timeout uint, cred map[string][]byte, db persistence.Database) {
 	credentials = cred
 	database = db
 	log.Printf("magnetico is ready to serve on %s!\n", address)
